pequeño refact
